LayoutTests/inspector/debugger/break-on-exception-throw-in-promise-with-catch-expected.txt

CONSOLE MESSAGE: line 44: testThrowingInPromiseWithCatch
CONSOLE MESSAGE: line 46: in promise
CONSOLE MESSAGE: line 7: throwing TestError
CONSOLE MESSAGE: line 49: in promise.catch
CONSOLE MESSAGE: line 50: DONE
Checking pause locations when pausing on all exceptions.

PAUSE #1 AT: doThrow:8:22
